requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
langchain-core>=0.3.21
langchain-google-genai>=2.0.7
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0
streamlit>=1.39.0
langchain-community>=0.3.3
langchain-huggingface>=0.2.0
huggingface-hub>=0.25.0
